requirements.txt
Flask
requests
Authlib
python-dotenv
gunicorn
